refactor ExcelTableHelper
